+++ canisters/audit/minimal_metrics/src/lib.rs
@@ -1,8 +1,0 @@
-//!
-//! Minimal metrics canister used for wasm-footprint auditing.
-//!
-
-icydb::start!();
-
-#[cfg(feature = "candid-export")]
-ic_cdk::export_candid!();