+++ crates/icydb/src/db/session/macros.rs
@@ -466,7 +466,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::require_one(&self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::require_one(&self.inner.execute_rows()?)
                 .map_err(Into::into)
         }
 
@@ -475,10 +475,8 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::require_some(
-                &self.inner.execute()?.into_rows()?,
-            )
-            .map_err(Into::into)
+            icydb_core::db::ResponseCardinalityExt::require_some(&self.inner.execute_rows()?)
+                .map_err(Into::into)
         }
 
         /// Materialize one row.
@@ -486,7 +484,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::row(self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::row(self.inner.execute_rows()?)
                 .map_err(Into::into)
         }
 
@@ -495,7 +493,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::try_row(self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::try_row(self.inner.execute_rows()?)
                 .map_err(Into::into)
         }
 
@@ -504,7 +502,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            Ok(self.inner.execute()?.into_rows()?.rows())
+            Ok(self.inner.execute_rows()?.rows())
         }
 
         /// Materialize an optional id.
@@ -512,13 +510,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            Ok(self
-                .inner
-                .execute()?
-                .into_rows()?
-                .iter()
-                .next()
-                .map(|row| row.id()))
+            Ok(self.inner.execute_rows()?.iter().next().map(|row| row.id()))
         }
 
         /// Materialize one required id.
@@ -526,7 +518,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::require_id(self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::require_id(self.inner.execute_rows()?)
                 .map_err(Into::into)
         }
 
@@ -535,7 +527,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::try_row(self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::try_row(self.inner.execute_rows()?)
                 .map(|row| row.map(|entry| entry.id()))
                 .map_err(Into::into)
         }
@@ -545,7 +537,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            Ok(self.inner.execute()?.into_rows()?.ids().collect())
+            Ok(self.inner.execute_rows()?.ids().collect())
         }
 
         /// Check whether an id is present in the response.
@@ -553,7 +545,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            Ok(self.inner.execute()?.into_rows()?.contains_id(id))
+            Ok(self.inner.execute_rows()?.contains_id(id))
         }
 
         /// Materialize one entity.
@@ -561,7 +553,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::entity(self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::entity(self.inner.execute_rows()?)
                 .map_err(Into::into)
         }
 
@@ -570,7 +562,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            icydb_core::db::ResponseCardinalityExt::try_entity(self.inner.execute()?.into_rows()?)
+            icydb_core::db::ResponseCardinalityExt::try_entity(self.inner.execute_rows()?)
                 .map_err(Into::into)
         }
 
@@ -579,7 +571,7 @@ macro_rules! impl_session_materialization_methods {
         where
             E: crate::traits::Entity,
         {
-            Ok(self.inner.execute()?.into_rows()?.entities())
+            Ok(self.inner.execute_rows()?.entities())
         }
     };
 }